Module sobol_burley::parts

The building blocks for making a Sobol sampler.

This module contains the internal components of the main samplers in this crate. You can use these to build alternative Sobol samplers. However, it is easy to mess things up in subtle ways. So unless you have unique requirements it is recommended to stick with the main samplers in the crate root.

Note: many of the functions in this module return reversed-bit integers, and take some of their parameters as reversed-bit integers as well. This is always indicated by a _rev postfix in the function name (for return values) and parameter names (for parameter values).

Examples

A simple, non-scrambled Sobol sequence function:

fn sobol(i: u32, dimension: u32) -> f32 {
    let sobol_int = sobol_rev(i.reverse_bits(), dimension).reverse_bits();

    u32_to_f32_norm(sobol_int)
}

A basic Owen-scrambled Sobol sequence function:

fn sobol_owen(i: u32, dimension: u32) -> f32 {
    let sobol_int_rev = sobol_rev(i.reverse_bits(), dimension);

    let sobol_owen_int = owen_scramble_rev(
        sobol_int_rev,
        hash(dimension),
    ).reverse_bits();

    u32_to_f32_norm(sobol_owen_int)
}

Structs

Int4

A packed set of four u32s.

Functions

hash

A fast 32-bit hash function.

hash_int4

Same as [hash_u32()] except on four numbers at once.

owen_scramble_int4_rev

Same as owen_scramble_rev(), except on 4 integers at a time.

owen_scramble_rev

Scramble n using a hash function that closely approximates a reverse-bit Owen scramble.

sobol_int4_rev

Same as sobol_rev() except returns four dimensions at once.

sobol_rev

Compute one dimension of a single sample in the Sobol sequence.

u32_to_f32_norm

Convert a u32 to a float in [0.0, 1.0).