Struct sobol_burley::parts::Int4

pub struct Int4 { /* private fields */ }

A packed set of four u32s.

Addition, subtraction, and multiplication are all wrapping.

Uses SIMD for computation on supported platforms.

Implementations

impl Int4

pub fn to_f32_norm(self) -> [f32; 4]

Convert each integer to a float in [0.0, 1.0).

Same behavior as parts::u32_to_f32_norm(), applied to each integer individually.

pub fn reverse_bits(self) -> Int4

Reverse the order of the bits in each integer.

Same behavior as reverse_bits() in the Rust standard library, applied to each integer individually.